[SERVER] update proto add version fields + update mapper (fix: incompatible type compare)

+++ DBSERVER/src/main/java/com/bht/parkingmap/dbserver/mapper/EnumMapper.java
@@ -19,8 +19,8 @@ import com.google.common.collect.HashBiMap;
 @Mapper(componentModel = "spring", nullValueMappingStrategy = NullValueMappingStrategy.RETURN_DEFAULT)
 public abstract class EnumMapper implements BaseBean {
 
-    private static final BiMap<Integer, UserRole> USER_ROLE_MAP = HashBiMap.create(5);
-    private static final BiMap<Integer, ParkingLotType> PARKING_LOT_TYPE_MAP = HashBiMap.create(4);
+    private static final BiMap<Long, UserRole> USER_ROLE_MAP = HashBiMap.create(5);
+    private static final BiMap<Long, ParkingLotType> PARKING_LOT_TYPE_MAP = HashBiMap.create(4);
 
 
     @Override
@@ -30,37 +30,37 @@ public abstract class EnumMapper implements BaseBean {
     }
 
     @Named("toUserRole")
-    UserRole toUserRole(Integer userRoleId) {
+    UserRole toUserRole(Long userRoleId) {
         return USER_ROLE_MAP.get(userRoleId);
     }
 
     @Named("toUserRoleId")
-    public Integer toUserRoleId(UserRole userRole) {
+    public Long toUserRoleId(UserRole userRole) {
         return USER_ROLE_MAP.inverse().get(userRole);
     }
 
     @Named("toParkingLotType")
-    ParkingLotType toParkingLotType(Integer parkingLotTypeId) {
+    ParkingLotType toParkingLotType(Long parkingLotTypeId) {
         return PARKING_LOT_TYPE_MAP.get(parkingLotTypeId);
     }
 
     @Named("toParkingLotId")
-    public Integer toParkingLotTypeId(ParkingLotType parkingLotType) {
+    public Long toParkingLotTypeId(ParkingLotType parkingLotType) {
         return PARKING_LOT_TYPE_MAP.inverse().get(parkingLotType);
     }
 
     private void initUserRoleMap() {
-        USER_ROLE_MAP.put(0, UserRole.UNRECOGNIZED);
-        USER_ROLE_MAP.put(1, UserRole.ADMIN);
-        USER_ROLE_MAP.put(2, UserRole.CUSTOMER);
-        USER_ROLE_MAP.put(3, UserRole.PARKING_LOT_EMPLOYEE);
-        USER_ROLE_MAP.put(4, UserRole.GOVERNMENT_EMPLOYEE);
+        USER_ROLE_MAP.put(0L, UserRole.UNRECOGNIZED);
+        USER_ROLE_MAP.put(1L, UserRole.ADMIN);
+        USER_ROLE_MAP.put(2L, UserRole.CUSTOMER);
+        USER_ROLE_MAP.put(3L, UserRole.PARKING_LOT_EMPLOYEE);
+        USER_ROLE_MAP.put(4L, UserRole.GOVERNMENT_EMPLOYEE);
     }
 
     private void initParkingLotTypeMap() {
-        PARKING_LOT_TYPE_MAP.put(0, ParkingLotType.UNRECOGNIZED);
-        PARKING_LOT_TYPE_MAP.put(1, ParkingLotType.PRIVATE);
-        PARKING_LOT_TYPE_MAP.put(2, ParkingLotType.BUILDING);
-        PARKING_LOT_TYPE_MAP.put(3, ParkingLotType.STREET);
+        PARKING_LOT_TYPE_MAP.put(0L, ParkingLotType.UNRECOGNIZED);
+        PARKING_LOT_TYPE_MAP.put(1L, ParkingLotType.PRIVATE);
+        PARKING_LOT_TYPE_MAP.put(2L, ParkingLotType.BUILDING);
+        PARKING_LOT_TYPE_MAP.put(3L, ParkingLotType.STREET);
     }
 }
